+++ pyam/timeseries.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
 
 import numpy as np
-import warnings
-
-from pyam.utils import (
-    isstr
-)
+from pyam.logger import logger
+from pyam.utils import isstr, cast_years_to_int
 
 # %%
+
 
 def fill_series(x, year):
     """Returns the value of a timeseries (indexed over years) for a year
@@ -50,18 +48,20 @@ def cumulative(x, first_year, last_year):
     last_year: int
         last year of the sum (inclusive)
     """
-
+    # if the timeseries does not cover the range `[first_year, last_year]`,
+    # return nan to avoid erroneous aggregation
     if min(x.index) > first_year:
-        # if the timeseries does not cover the range `[first_year, last_year]`,
-        # replace by nan to avoid erroneous aggregation
-        warnings.warn('timeseries {} does not start by {}'.format(x.name,
-                      first_year))
+        logger().warning('the timeseries `{}` does not start by {}'.format(
+            x.name or x, first_year))
+        return np.nan
+    if max(x.index) < last_year:
+        logger().warning('the timeseries `{}` does not extend until {}'
+                         .format(x.name or x, last_year))
         return np.nan
 
-    if max(x.index) < last_year:
-        warnings.warn('the timeseries {} does not extend until {}'
-                      .format(x.name, last_year))
-        return np.nan
+    # cast tiemseries colums to `int` if necessary
+    if not x.index.dtype == 'int64':
+        cast_years_to_int(x, index=True)
 
     x[first_year] = fill_series(x, first_year)
     x[last_year] = fill_series(x, last_year)